+++ tests/test_compat.py
@@ -5,8 +5,9 @@ import pytest
 
 from anyio import (
     TaskInfo, create_capacity_limiter, create_condition, create_event, create_lock,
-    create_semaphore, create_task_group, fail_after, get_current_task, get_running_tasks,
-    maybe_async, maybe_async_cm, move_on_after, open_cancel_scope, open_signal_receiver, sleep)
+    create_memory_object_stream, create_semaphore, create_task_group, fail_after, get_current_task,
+    get_running_tasks, maybe_async, maybe_async_cm, move_on_after, open_cancel_scope,
+    open_signal_receiver, sleep)
 
 pytestmark = pytest.mark.anyio
 
@@ -80,6 +81,11 @@ class TestDeprecations:
         lock.acquire_nowait()
         with pytest.deprecated_call():
             await lock.release()
+
+    async def test_memory_object_stream_send_nowait(self):
+        send, receive = create_memory_object_stream(1)
+        with pytest.deprecated_call():
+            await send.send_nowait(None)
 
     async def test_semaphore_release(self):
         semaphore = create_semaphore(1)